Restrict shuffle to positive weights
For one, kind of a stupid thing in the first place to have the zero
weight cases. A lot of unnecessary computation for something that
can just be avoided by not passing zero weights. In any case,
turns out it also didn't work properly anyway, so instead of trying
to fix it, just deleted the nonsense.